feat: implement Flask API for Booking.com header extraction with endpoints for health check and header retrieval
