+++ auth_headers/tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from components.api import app
+
+@pytest.fixture
+def client():
+    """Create a test client for the Flask app."""
+    app.config["TESTING"] = True
+    with app.test_client() as client:
+        yield client
